refactor: apply eslint app+components
