+++ app/src/app/(account)/submitted-trips.tsx
@@ -1,4 +1,3 @@
-import React, { useEffect, useState } from "react";
 import {
   SafeAreaView,
   View,
