+++ app/src/contexts/TripSearchContext.tsx
